Rearranged filestructure; added logo assets; changed applications icons; Renamed namespaces according to their folder; added functions in package.appxmanifest; added services in package.appxmanifest;
