+++ Pr0gramm/API/Image.cs
@@ -1,11 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using asapJson;
-
-namespace Pr0gramm.pr0
+﻿namespace Pr0gramm.API
 {
     public class Image
     {
