+++ Pr0gramm/API/User.cs
@@ -1,11 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Windows.UI.Xaml.Media;
 
-namespace Pr0gramm.pr0
+namespace Pr0gramm.API
 {
     public class User
     {
