+++ Pr0gramm/MainPage.xaml
@@ -4,7 +4,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:Pr0gramm"
     xmlns:pages="using:Pr0gramm.Pages"
-    xmlns:uc="using:Pr0gramm.UserControls"
+    xmlns:uc="using:Pr0gramm.UI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d" Height="Auto" Width="Auto">
