+++ Pr0gramm/Pr0gramm.csproj
@@ -102,24 +102,24 @@
     <Compile Include="MainPage.xaml.cs">
       <DependentUpon>MainPage.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Pages\ContentPresenter.xaml.cs">
+    <Compile Include="UI\Pages\ContentPresenter.xaml.cs">
       <DependentUpon>ContentPresenter.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Pages\ImageView.xaml.cs">
+    <Compile Include="UI\Pages\ImageView.xaml.cs">
       <DependentUpon>ImageView.xaml</DependentUpon>
     </Compile>
-    <Compile Include="pr0\Image.cs" />
-    <Compile Include="pr0\ItemInfo.cs" />
-    <Compile Include="pr0\User.cs" />
+    <Compile Include="API\Image.cs" />
+    <Compile Include="API\ItemInfo.cs" />
+    <Compile Include="API\User.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Settings.cs" />
-    <Compile Include="UserControls\Pr0Button.xaml.cs">
+    <Compile Include="UI\Controls\Pr0Button.xaml.cs">
       <DependentUpon>Pr0Button.xaml</DependentUpon>
     </Compile>
-    <Compile Include="UserControls\Tag.xaml.cs">
+    <Compile Include="UI\Controls\Tag.xaml.cs">
       <DependentUpon>Tag.xaml</DependentUpon>
     </Compile>
-    <Compile Include="UserControls\UserComment.xaml.cs">
+    <Compile Include="UI\Controls\UserComment.xaml.cs">
       <DependentUpon>UserComment.xaml</DependentUpon>
     </Compile>
   </ItemGroup>
@@ -130,13 +130,36 @@
     <None Include="Pr0gramm_TemporaryKey.pfx" />
   </ItemGroup>
   <ItemGroup>
+    <Content Include="Assets\BadgeLogo.scale-100.png" />
+    <Content Include="Assets\BadgeLogo.scale-125.png" />
+    <Content Include="Assets\BadgeLogo.scale-150.png" />
+    <Content Include="Assets\BadgeLogo.scale-200.png" />
+    <Content Include="Assets\BadgeLogo.scale-400.png" />
+    <Content Include="Assets\SplashScreen.scale-100.png" />
+    <Content Include="Assets\SplashScreen.scale-125.png" />
+    <Content Include="Assets\SplashScreen.scale-150.png" />
+    <Content Include="Assets\SplashScreen.scale-400.png" />
+    <Content Include="Assets\Square100x100Logo.scale-200.png" />
+    <Content Include="Assets\Square150x150Logo.scale-100.png" />
+    <Content Include="Assets\Square150x150Logo.scale-400.png" />
+    <Content Include="Assets\Square200x200Logo.scale-100.png" />
+    <Content Include="Assets\Square200x200Logo.scale-125.png" />
+    <Content Include="Assets\Square200x200Logo.scale-150.png" />
+    <Content Include="Assets\Square200x200Logo.scale-200.png" />
+    <Content Include="Assets\Square200x200Logo.scale-400.png" />
+    <Content Include="Assets\Square310x310Logo.scale-400.png" />
+    <Content Include="Assets\Square44x44Logo.scale-400.png" />
+    <Content Include="Assets\Square50x50Logo.scale-100.png" />
+    <Content Include="Assets\Square63x63Logo.scale-125.png" />
+    <Content Include="Assets\Square71x71Logo.scale-400.png" />
+    <Content Include="Assets\Square75x75Logo.scale-150.png" />
+    <Content Include="Assets\Wide310x150Logo.scale-400.png" />
     <Content Include="Properties\Default.rd.xml" />
     <Content Include="Assets\LockScreenLogo.scale-200.png" />
     <Content Include="Assets\SplashScreen.scale-200.png" />
     <Content Include="Assets\Square150x150Logo.scale-200.png" />
     <Content Include="Assets\Square44x44Logo.scale-200.png" />
     <Content Include="Assets\Square44x44Logo.targetsize-24_altform-unplated.png" />
-    <Content Include="Assets\StoreLogo.png" />
     <Content Include="Assets\Wide310x150Logo.scale-200.png" />
   </ItemGroup>
   <ItemGroup>
@@ -148,11 +171,11 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-    <Page Include="Pages\ContentPresenter.xaml">
+    <Page Include="UI\Pages\ContentPresenter.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Pages\ImageView.xaml">
+    <Page Include="UI\Pages\ImageView.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -161,15 +184,15 @@
       <Generator>MSBuild:Compile</Generator>
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </Page>
-    <Page Include="UserControls\Pr0Button.xaml">
+    <Page Include="UI\Controls\Pr0Button.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="UserControls\Tag.xaml">
+    <Page Include="UI\Controls\Tag.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="UserControls\UserComment.xaml">
+    <Page Include="UI\Controls\UserComment.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
