+++ Pr0gramm/UI/Controls/Pr0Button.xaml
@@ -1,8 +1,8 @@
 ﻿<UserControl
-    x:Class="Pr0gramm.UserControls.Pr0Button"
+    x:Class="Pr0gramm.UI.Controls.Pr0Button"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Pr0gramm.UserControls"
+    xmlns:local="using:Pr0gramm.UI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d"
