+++ Pr0gramm/UI/Pages/ContentPresenter.xaml
@@ -1,8 +1,8 @@
 ﻿<Page
-    x:Class="Pr0gramm.Pages.ContentPresenter"
+    x:Class="Pr0gramm.UI.Pages.ContentPresenter"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Pr0gramm"
+    xmlns:local="using:Pr0gramm.UI.Pages"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
