+++ Pr0gramm/UI/Pages/ContentPresenter.xaml.cs
@@ -1,28 +1,17 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Runtime.InteropServices.WindowsRuntime;
-using Windows.Foundation;
-using Windows.Foundation.Collections;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
-using Windows.UI.Xaml.Controls.Primitives;
-using Windows.UI.Xaml.Data;
-using Windows.UI.Xaml.Input;
-using Windows.UI.Xaml.Media;
 using Windows.UI.Xaml.Navigation;
 using System.Net;
-using System.ComponentModel;
-using System.Text;
 using asapJson;
 using System.Threading.Tasks;
-using Windows.UI.Notifications;
 
 
 // The Blank Page item template is documented at http://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace Pr0gramm.Pages
+namespace Pr0gramm.UI.Pages
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
@@ -116,7 +105,7 @@ namespace Pr0gramm.Pages
         }
 
         public ViewSource Source { get; set; }
-        private List<pr0.Image> images;
+        private List<API.Image> images;
         private int newestItem;
         private int oldestItem;
 
@@ -125,7 +114,7 @@ namespace Pr0gramm.Pages
             this.InitializeComponent();
             newestItem = -1;
             oldestItem = -1;
-            images = new List<pr0.Image>();
+            images = new List<API.Image>();
         }
         protected async override void OnNavigatedTo(NavigationEventArgs e)
         {
@@ -182,10 +171,10 @@ namespace Pr0gramm.Pages
                 this.oldestItem = (int)item["promoted"].getValue_Number();
             else
                 this.oldestItem = (int)item["id"].getValue_Number();
-            List<pr0.Image> imgList = new List<pr0.Image>();
+            List<API.Image> imgList = new List<API.Image>();
             foreach(var it in items)
             {
-                imgList.Add(new pr0.Image(it));
+                imgList.Add(new API.Image(it));
             }
             if (fd == FetchDirection.Newer)
                 this.images.InsertRange(0, imgList);
