+++ Pr0gramm/UI/Pages/ImageView.xaml
@@ -1,9 +1,9 @@
 ﻿<Page
-    x:Class="Pr0gramm.Pages.ImageView"
+    x:Class="Pr0gramm.UI.Pages.ImageView"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Pr0gramm.Pages"
-    xmlns:uc="using:Pr0gramm.UserControls"
+    xmlns:local="using:Pr0gramm.UI.Pages"
+    xmlns:uc="using:Pr0gramm.UI.Controls"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d" Width="Auto">
